adding 104-fibonacci.c file

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,13 +8,13 @@
 int main(void)
 {
 int i;
-long long int first = 1, second = 2, next;
+long int first = 1, second = 2, next;
 
-printf("%lld, %lld, ", first, second);
+printf("%ld, %ld, ", first, second);
 
 for (i = 3; i <= 98; i++) {
 next = first + second;
-printf("%lld", next);
+printf("%ld", next);
 
 if (i != 98)
 printf(", ");
